% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adjust_for_collision}
\alias{adjust_for_collision}
\title{Adjust for collision}
\usage{
adjust_for_collision(input_gr)
}
\arguments{
\item{input_gr}{A GRanges with 1 meta: 'old_width'}
}
\value{
A list with:
        - 'pot_issues': A dataframe with exons which overlaps between
                       `input_gr` and itself while gene_ids are different
        - 'new_gr': A GRanges identical to `input_gr` except that start/end
                    have been adjusted to prevent collisions.
}
\description{
A function that from a GRanges with 'old_width'
Change the starts and ends to prevent collisions larger than
with old coordinates
}
