% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_coo_changes}
\alias{apply_coo_changes}
\title{Apply start end changes}
\usage{
apply_coo_changes(input_gr)
}
\arguments{
\item{input_gr}{A GRanges with 2 metas: 'new_start' and 'new_end'}
}
\value{
A GRanges identical to `input_gr` except that
        start and end are now new_start new_end and
        initial start and end have been stored into old_start and old_end
}
\description{
A function that from a GRanges with 'new_start' and 'new_end'
Change the starts and ends
and store old starts/ends into 'old_start', 'old_end'.
}
