% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{debug_msg}
\alias{debug_msg}
\title{Display debug messages if verbose allows it}
\usage{
debug_msg(message = NULL, ...)
}
\arguments{
\item{message}{String to display}

\item{...}{Other parameters for rlang::inform}
}
\value{
Nothing
}
\description{
A function that extend rlang::inform
to display a message if the verbose is at "debug"
and show content of the variable
}
