% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_last_exons}
\alias{extract_last_exons}
\title{Extract last exons}
\usage{
extract_last_exons(
  input_gr,
  grouping_variable = "transcript_id",
  invert = FALSE
)
}
\arguments{
\item{input_gr}{A GRanges from a gtf}

\item{grouping_variable}{A string with the name of the metadata
which should be used to group}

\item{invert}{A boolean that indicates if you want all except the last exons}
}
\value{
A GRanges which contains a subset of `input_gr`
}
\description{
A function that from a GRanges from gtf
select only entries for the last exons
If multiple exons overlap the last base of the grouping_variable,
they will all be reported.
}
