% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_new_exons}
\alias{filter_new_exons}
\title{Filter new exons for collision}
\usage{
filter_new_exons(all_exons_interesting, input_gr_to_extend)
}
\arguments{
\item{all_exons_interesting}{A GRanges with exons to trim and filter}

\item{input_gr_to_extend}{A GRanges to overlap}
}
\value{
A GRanges subset of `all_exons_interesting`
}
\description{
A function that from 2 GRanges filter exons from the first one
so they do not go three prime to the first collision with the second one.
}
