% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{five_prime_pos}
\alias{five_prime_pos}
\title{Get five prime position}
\usage{
five_prime_pos(input_gr)
}
\arguments{
\item{input_gr}{A GRanges or GRangeList}
}
\value{
A vector of integers
}
\description{
A function that from a GRanges
gives the 5' position
}
