% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_extending_overlap}
\alias{get_extending_overlap}
\title{Overlap exons and filter for potential extension}
\usage{
get_extending_overlap(input_gr_to_extend, input_gr_to_overlap, verbose = 1)
}
\arguments{
\item{input_gr_to_extend}{A GRanges with exons to extend
(strand * are excluded)}

\item{input_gr_to_overlap}{A GRanges with intervals to overlap}

\item{verbose}{An integer that indicates the level of verbosity
0 = silent, 1 = statistics, 2 = progression.}
}
\value{
A GRanges which is a subset of `input_gr_to_extend` where
        new_start and new_end have been added to match the 3' end of
        `input_gr_to_overlap` if they overlap
}
\description{
A function that from 2 GRanges returns a dataframe describing
overlaps that could lead to extension of the first GRanges
}
