% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{overlap_different_genes}
\alias{overlap_different_genes}
\title{Get overlaps from different genes}
\usage{
overlap_different_genes(gr1, gr2)
}
\arguments{
\item{gr1}{A GRanges with 'gene_id'}

\item{gr2}{A GRanges with 'gene_id'}
}
\value{
a data.frame with overlaps between gr1 and gr2 when gene_id from gr1
        is different from gene_id from gr2. The data.frame has 4 columns:
        `queryHits`, `subjectHits`, `query_gene_id` and `subject_gene_id`
}
\description{
A function that from 2 GRanges generates a dataframe
With queryHits, subjectHits when the gene_id is different
}
