% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{progression_msg}
\alias{progression_msg}
\title{Display progression messages if verbose allows it}
\usage{
progression_msg(...)
}
\arguments{
\item{...}{Parameters for rlang::inform}
}
\value{
Nothing
}
\description{
A function that extend rlang::inform
to display a message if the verbose is at "debug" or "progression"
}
