% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{three_prime_pos}
\alias{three_prime_pos}
\title{Get three prime position}
\usage{
three_prime_pos(input_gr)
}
\arguments{
\item{input_gr}{A GRanges or GRangeList}
}
\value{
A vector of integers
}
\description{
A function that from a GRanges
gives the 3' position
}
